Rewrite DateField
